Added simple tests for the Election Module
I also had to run some debugging because the tests for this module kept failing
from time to time, with what seemed like no explanation.
Turns out there was some problems with pointers and stuff.
I'll take note to run Valgrind or similar when the progress is in good shape.